+++ db/queries/comment.js
@@ -21,15 +21,27 @@ export async function findAll(postId) {
     console.log("COMMENT FINDALL");
     return await Comments.findAll({
         where: { postId },
+        order: [["createdAt", "DESC"], ["commentId", "DESC"]],
         attributes: {
             exclude: ["postId", "deletedAt"]
         },
-        // include: [UserModel]
+        include: {
+            model: Users,
+            attributes: ["nickname"]
+        }
     });
 }
 
 export async function updateOne(comment) {
     console.log("COMMENT UPDATEONE");
+
+    const check = await Comments.findOne({
+        where: { commentId: comment.commentId, userId: comment.userId }
+    });
+    if (check === null) {
+        return [ null ];
+    }
+
     return await Comments.update({
         comment: comment.comment
     }, {
@@ -37,9 +49,17 @@ export async function updateOne(comment) {
     });
 }
 
-export async function deleteOne(commentId) {
+export async function deleteOne(ids) {
     console.log("COMMENT DELETEONE");
+
+    const check = await Comments.findOne({
+        where: { commentId: ids.commentId, userId: ids.userId }
+    });
+    if (check === null) {
+        return [ null ];
+    }
+
     return await Comments.destroy({
-        where: { commentId }
+        where: { commentId: ids.commentId }
     });
 }
